Improved onReceive() method.
Dropped the didReceiveFinalMessage() delegate method and added
isMessageComplete parameter to didReceiveData().

In onReceive(), renamed isComplete parameter to isMessageComplete to
better describe what it represents.  Also added check for nil Data
parameter that indicates the end of the message.  Make sure to only
trigger the next receive() call when more data is expected.